feat: 공개 글 탐색 경로 추가
